u-boot: Add usb support for the rpi4 and generic ARM64
This commit changes the bcm-2xxx-rpi4.conf to use a different u-boot
machine that is generic for all raspberry pi 64 configurations.  This
will allow the BSP to boot on teh rpi3 and rpi4.

This commit also adds the USB support for the rpi4 which will allow a
keyboard to be used with u-boot when the HDMI port is plugged in.  It
also enables the ability to boot off a USB block storage device.

[ Issue: LINUXEXEC-8353 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>