u-boot: Build the u-boot binary for QEMU
This change will build two u-boot binaries instead of one anytime that u-boot is built.

The second binary is named: qemu-u-boot-bcm-2xxx-rpi4.bin

The runqemu requires an image to have both the keyword "-image-" and
end in ".hddimg" in order to properly setup the environment to use u-boot.

This is how to create an image using bootfs.sh for example which will boot:

  ../layers/wr-ostree/scripts/bootfs.sh -e core-image-minimal -s 0 -l 0 -n
  qemu-img create -f raw boot-image-qemu.hddimg 10G
  dd if=ustart.img of=boot-image-qemu.hddimg conv=notrunc
  runqemu nographic slirp \
    qemuparams="-m 1024 -bios tmp-glibc/deploy/images/bcm-2xxx-rpi4/qemu-u-boot-bcm-2xxx-rpi4.bin" \
    boot-image-qemu.hddimg

[ Issue: LINUXEXEC-8353 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>