rpi-graphics: remove faad2 dependency
Issue: LINCD-2784

faad library has a commercial license, it should only be included when
the commercial license is supported. This is the backport of the
solution provided by the meta-raspberrypi upstream (c5bc55687323a703aa21a0170f6509a1b04bd78b).

Since we don't support the commercial license, faad should never be
included, hence shouldn't be added to the PNWHITELIST.

Signed-off-by: Mingde (Matthew) Zeng <matthew.zeng@windriver.com>