Nail down the kernel version in machine conf
Issue: LINCD-3001

In some cases, when the poky bump to a new version kernel, but we don't
want to do that for our own BSPs. This will cause build failure due to
the setting of "PREFERRED_VERSION_linux-yocto" in meta-poky/conf/distro/poky.conf
when building a poky distro with our own BSPs. So explicitly nail down
the kernel version in machine conf to fix this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>