bcm-2xxx-rpi: remove the temprorary patches for linux-ycoto-dev kernel
Issue: LINCD-3259

There are 2 temprorary patches for former linux-ycoto-dev
kernel, v5.8. But now, they are useless for the v5.9 linux-ycoto-dev
kernel.

Detail removing reason as below:
0001-arch-arm-dts-remove-unused-power-supply-dts-property.patch
It is used to remove vmmc-supply property for SD device, because the
fixed regurator is built as module, it is not able to be used when kernel
boots up. But now with latest kernel v5.9, fixed regurator is built into
kernel image. So, fixed regurator is able to be used during kernel boots
up and it is not need to remove the vmmc-supply property anymore.

0001-driver-usb-host-skip-vl805-firmware-init-for-raspber.patch
It is a wrokaound for USB feature on raspberry pi4 rev1.4, 8GB DDR
memory board. But now, latest sdk patches has fixed this issue, and
USB feature works fine with v5.9 kernel.

Signed-off-by: Meng Li <Meng.Li@windriver.com>