bcm-2xxx-rpi: set correct boardflags3 parameter for raspberrypi pi4 B platform
Issue: LINCD-4966

This issue is introduced by linux-firmware upstream commit
870b805dc77e ("brcm: Update Raspberry Pi 3B+/4B NVRAM for
downstream changes"). Using the latest linux-firmware package,
the boardflags3 parameter to 0x48200100 which indicates the
firmware that an external LPO is present. But on the RPi4 no external
LPO exists.So, correct it to support WIFI feature on this platform.

Upstream-Status: Backport

Signed-off-by: Meng Li <Meng.Li@windriver.com>