bcm-2xxx-rpi4.conf: Clear DEFAULT_DTB for ostree
Issue: LINCD-5223

Don't set a default dtb for the board, the u-boot loads it into the fdt_addr
automatically, a lot of devices would not work if we load it again to override
fdt_addr.

The u-boot-uenv has been changed to load the fdt_file when it is set, so we
need clear DEFAULT_DTB which is fdt_file to make the board work, the
DEFAULT_DTB is only used by u-boot-uenv.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>